# examples/api-platform/openapi.yaml
openapi: 3.0.3
info:
  title: API Platform
  description: |
    A production-ready API Platform template demonstrating modern API-first product development.

    ## Authentication

    This API uses API Key authentication. Include your API key in the `X-API-Key` header:

    ```
    X-API-Key: pk_live_xxxxxxxxxxxx
    ```

    ## Rate Limiting

    API requests are rate limited based on your plan:
    - **Free**: 100 requests/hour
    - **Pro**: 1,000 requests/hour
    - **Enterprise**: 10,000 requests/hour

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

    ## Versioning

    The API is versioned via URL path. Current versions:
    - `v1`: Stable version
    - `v2`: Latest version (may have breaking changes)

    ## Webhooks

    Register webhook endpoints to receive event notifications. Events include:
    - `resource.created`
    - `resource.updated`
    - `resource.deleted`
    - `api_key.created`
    - `api_key.deleted`
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
    url: https://developer.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000/v1
    description: Local development server
  - url: https://api.example.com/v1
    description: Production server

security:
  - ApiKeyAuth: []

tags:
  - name: Health
    description: Health check endpoints
  - name: API Keys
    description: Manage API keys
  - name: Resources
    description: CRUD operations for resources
  - name: Webhooks
    description: Webhook endpoint management
  - name: Analytics
    description: Usage analytics and metrics

paths:
  /health:
    get:
      summary: Health check
      description: Check if the API is running and healthy
      tags:
        - Health
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: healthy
                version: 1.0.0
                timestamp: '2024-01-15T10:30:00Z'
                uptime: 86400

  /health/ready:
    get:
      summary: Readiness probe
      description: Check if the API is ready to accept requests
      tags:
        - Health
      security: []
      responses:
        '200':
          description: API is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  ready:
                    type: boolean
                  checks:
                    type: object

  /api-keys:
    get:
      summary: List API keys
      description: Get all API keys for the authenticated user
      tags:
        - API Keys
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create API key
      description: Generate a new API key
      tags:
        - API Keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
            example:
              name: Production Key
              tier: pro
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyWithSecret'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'

  /api-keys/{id}:
    get:
      summary: Get API key
      description: Get details of a specific API key
      tags:
        - API Keys
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Revoke API key
      description: Permanently revoke an API key
      tags:
        - API Keys
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: API key revoked successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api-keys/{id}/rotate:
    post:
      summary: Rotate API key
      description: Generate a new API key secret while keeping the same ID
      tags:
        - API Keys
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key rotated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyWithSecret'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /resources:
    get:
      summary: List resources
      description: Get all resources with pagination
      tags:
        - Resources
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: cursor
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, archived]
      responses:
        '200':
          description: List of resources
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Resource'
                  pagination:
                    $ref: '#/components/schemas/CursorPagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'

    post:
      summary: Create resource
      description: Create a new resource
      tags:
        - Resources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRequest'
            example:
              name: My Resource
              description: A sample resource
              metadata:
                key: value
      responses:
        '201':
          description: Resource created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'

  /resources/{id}:
    get:
      summary: Get resource
      description: Get details of a specific resource
      tags:
        - Resources
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resource details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Update resource
      description: Update an existing resource
      tags:
        - Resources
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResourceRequest'
      responses:
        '200':
          description: Resource updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete resource
      description: Permanently delete a resource
      tags:
        - Resources
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Resource deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /webhook-endpoints:
    get:
      summary: List webhook endpoints
      description: Get all registered webhook endpoints
      tags:
        - Webhooks
      responses:
        '200':
          description: List of webhook endpoints
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookEndpoint'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create webhook endpoint
      description: Register a new webhook endpoint
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
            example:
              url: https://example.com/webhooks
              events:
                - resource.created
                - resource.updated
              secret: whsec_xxxxxxxxxxxx
      responses:
        '201':
          description: Webhook endpoint created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /webhook-endpoints/{id}:
    get:
      summary: Get webhook endpoint
      description: Get details of a specific webhook endpoint
      tags:
        - Webhooks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook endpoint details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete webhook endpoint
      description: Remove a webhook endpoint
      tags:
        - Webhooks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Webhook endpoint deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /webhook-endpoints/{id}/test:
    post:
      summary: Test webhook endpoint
      description: Send a test event to the webhook endpoint
      tags:
        - Webhooks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test webhook sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookDelivery'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /webhook-deliveries:
    get:
      summary: List webhook deliveries
      description: Get delivery history for webhooks
      tags:
        - Webhooks
      parameters:
        - name: endpoint_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, delivered, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of webhook deliveries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookDelivery'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /analytics/usage:
    get:
      summary: Get usage analytics
      description: Get API usage statistics
      tags:
        - Analytics
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Usage analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageAnalytics'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        version:
          type: string
        timestamp:
          type: string
          format: date-time
        uptime:
          type: number
          description: Uptime in seconds

    ApiKey:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
        tier:
          type: string
          enum: [free, pro, enterprise]
        status:
          type: string
          enum: [active, revoked]
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true

    ApiKeyWithSecret:
      allOf:
        - $ref: '#/components/schemas/ApiKey'
        - type: object
          properties:
            secret:
              type: string
              description: Full API key secret (only shown once on creation)

    CreateApiKeyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        tier:
          type: string
          enum: [free, pro, enterprise]
          default: free

    Resource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, archived]
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateResourceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        metadata:
          type: object
          additionalProperties: true

    UpdateResourceRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [active, inactive, archived]
        metadata:
          type: object
          additionalProperties: true

    WebhookEndpoint:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, disabled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateWebhookRequest:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              [
                resource.created,
                resource.updated,
                resource.deleted,
                api_key.created,
                api_key.deleted,
              ]
        secret:
          type: string
          description: Optional secret for webhook signature verification

    WebhookDelivery:
      type: object
      properties:
        id:
          type: string
        endpoint_id:
          type: string
        event_id:
          type: string
        status:
          type: string
          enum: [pending, delivered, failed]
        http_status:
          type: integer
          nullable: true
        delivered_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
        retry_count:
          type: integer

    WebhookEvent:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        created_at:
          type: string
          format: date-time
        data:
          type: object

    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        total_pages:
          type: integer

    CursorPagination:
      type: object
      properties:
        has_more:
          type: boolean
        next_cursor:
          type: string
          nullable: true
        prev_cursor:
          type: string
          nullable: true

    UsageAnalytics:
      type: object
      properties:
        total_requests:
          type: integer
        successful_requests:
          type: integer
        failed_requests:
          type: integer
        requests_by_endpoint:
          type: array
          items:
            type: object
            properties:
              endpoint:
                type: string
              count:
                type: integer
        requests_over_time:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: unauthorized
              message: API key is missing or invalid

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: not_found
              message: The requested resource was not found

    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: validation_error
              message: The request data is invalid
              details:
                - field: name
                  message: Name is required

    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: rate_limit_exceeded
              message: Rate limit exceeded. Please try again later.